Barebones co-op playstyle implemented using turns + Multiple fixes to the code
Turn based co-op implemented, however still require a win/lose/draw condition to be added with the checkmate/stalemate check during each turn, in addition to a timer for the players.

Unnecessary printfs commented out for a much clear output, in addition to correcting grammatical and logical errors in comments (XNOR was discussed when XOR was supposed to be used in opposing chess pieces' collision) and changing the inequality to an XOR bitwise operator for the collision check between two opposing pieces.

Error relating to freeing the allocated memory in next_move fixed as a consequence of fixing check_move's manipulation of the passed strings.

TO-DO List updated with the remaining tasks